+++ INaturalistIOS/ObservationsViewController.h
@@ -17,10 +17,10 @@
 
 
 @interface ObservationsViewController : UITableViewController <ObservationDetailViewControllerDelegate, UIAlertViewDelegate, SyncQueueDelegate, RKObjectLoaderDelegate, RKRequestDelegate, RKObjectMapperDelegate>
-@property (nonatomic, assign) int observationsToSyncCount;
-@property (nonatomic, assign) int syncedObservationsCount;
-@property (nonatomic, assign) int observationPhotosToSyncCount;
-@property (nonatomic, assign) int syncedObservationPhotosCount;
+@property (nonatomic, assign) NSInteger observationsToSyncCount;
+@property (nonatomic, assign) NSInteger syncedObservationsCount;
+@property (nonatomic, assign) NSInteger observationPhotosToSyncCount;
+@property (nonatomic, assign) NSInteger syncedObservationPhotosCount;
 @property (nonatomic, strong) NSArray *syncToolbarItems;
 @property (strong, nonatomic) IBOutlet UIBarButtonItem *syncButton; // if the button is just kind of floating and not assigned a super view, it will get deallocated UNLESS we have a strong reference here
 @property (strong, nonatomic) IBOutlet UIBarButtonItem *editButton;
@@ -38,7 +38,7 @@
 - (void)loadData;
 - (void)reload;
 - (void)checkSyncStatus;
-- (int)itemsToSyncCount;
+- (NSInteger)itemsToSyncCount;
 - (void)clickedActivity:(id)sender event:(UIEvent *)event;
 - (void)handleNSManagedObjectContextDidSaveNotification:(NSNotification *)notification;
 - (BOOL)isSyncing;
